services/notification-service/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
httpx==0.27.0
python-dotenv==1.0.1
pydantic-settings==2.2.1
prometheus-fastapi-instrumentator==7.0.0
